Makes DataSetView available for QMLComponent
